DQ-294 - Extract postMessage logic to usePostMessageSeriesSwitching hook
